Movement works. Size problem solved.
Multiple vehicles can be placed by the user and correctly move around
the circuit in unison.
The problem that the model was too small to see has been fixed by simply
inputting a smaller dataset - if given the time add a feature that adds
a line sticking out from each vehicles location if the dataset is larger
than a certain size.