Traffic lights work!
Traffic light logic added and they seem to work. There is a bug where it
sometimes takes the user to click start twice to start. Next step is
adding more user input to the control panel and adding the statistics.
If there is time add vehicle acceleration and max speeds.